organized components folder with subfolders for components with subcomponents for easier file managment
