+++ src/assets/socialIcons.js/index.js
@@ -1,0 +1,5 @@
+import Instagram from "./Instagram";
+import LinkedIn from './LinkedIn'
+import GitHub from "./GitHub";
+
+export {Instagram, LinkedIn, GitHub}
