+++ src/components/About/aboutCopy.js
@@ -1,0 +1,3 @@
+const bio = "Eddie Lopez (Lalo) is a Chicano illustrator and web developer based in the San Joaquin valley. After studying illustration at the California College of the Arts in San Francisco, he went on to attend Lambda School’s fullstack web development program. Branding himself as both a creative and a developer, his ultimate goal is to work on projects where he can combine the two fields to create web solutions that are beautifully designed and impactful for users."
+
+export default bio
