+++ src/components/Gallery/Gallery.js
