+++ src/components/Gallery/QuickGallery.js
