+++ src/components/Gallery/index.js
@@ -1,0 +1,4 @@
+import Gallery from "./Gallery";
+import QuickGallery from './QuickGallery'
+
+export {Gallery,QuickGallery}
