+++ src/components/styled-components/index.js
@@ -2,5 +2,16 @@ import AppWrapper from "./AppWrapper";
 import Section from "./Section";
 import { Input as StyledInput, Label, TextArea, Form } from "./Forms";
 import { Nav, NavLinks } from "./Navigation";
+import Button from "./Button";
 
-export { AppWrapper, StyledInput, Label, TextArea, Form, Section, Nav, NavLinks};
+export {
+  AppWrapper,
+  StyledInput,
+  Label,
+  TextArea,
+  Form,
+  Section,
+  Nav,
+  NavLinks,
+  Button,
+};
